+++ lib/domain/index.ts
@@ -2,3 +2,4 @@ export * from "./growth";
 export * from "./measurement";
 export * from "./pet";
 export * from "./units";
+export * from "./vaccination";